+++ superloc/manager/templates/manager/vehicles_availability_agency_choice.html
@@ -8,7 +8,31 @@
   <form action="{% url 'manager:vehicles_availability_agency_choice' %}" method="post">
     {%csrf_token%}
     {{ form.as_p }}
-    <input type="submit" value="Vérifier les disponibilités">
+    <div>
+      <p>
+        <label for="category">Catégorie de véhicule :</label>
+        <select id="category" required>
+        <option value="L">L - Compacte automatique</option>
+        </select>
+      </p>
+    </div>
+    <div>
+      <p>
+        <label for="date_departure">Départ :</label>
+        <input type="datetime-local" id="date_departure"
+               name="date_start"
+               min="2022-01-01T00:00" max="2022-12-01T00:00" required>
+      </p>
+    </div>
+    <div>
+      <p>
+        <label for="date_back">Retour :</label>
+        <input type="datetime-local" id="date_back"
+               name="date_back"
+               min="2022-01-01T00:00" max="2022-12-01T00:00" required>
+      </p>
+    </div>
+    <p><input type="submit" value="Vérifier les disponibilités"></p>
   </form>
 </div>
 {% endblock %}
