+++ superloc/visitor/models.py
@@ -100,3 +100,115 @@ class Booking(models.Model):
     customer = models.ForeignKey(Customer, on_delete=models.PROTECT)
     date_start = models.DateTimeField('Date de début de location')
     date_end = models.DateTimeField('Date de fin de location')
+
+def vehicle_availability_list(category, agency_departure, date_departure, date_back):
+    bookings_to_exclude_list = (
+        Bookings.objects.all()
+            .filter(agency=agency_departure)
+            .filter(date_end__gte=date_departure, date_start__lte=date_departure)
+            .filter(date_start__lte=date_back, date_end__gte=date_start)
+    )
+    available_vehicles = (
+        Vehicle.objects.all()
+            .filter(is_active=True)
+            .filter(agency=agency)
+            .filter(category=category)
+            .exclude(id__in=Subquery(bookings_to_exclude_list.values('vehicle')))
+    )
+    return available_vehicles
+
+'''
+.exclude(id__in=Subquery(liste_contrats_conclus_bloquants_sur_la_periode_demandee.values('vehicule'))) #à condition que le véhicule ne figure pas dans un contrat déjà conclu aux dates souhaitées​
+    )​
+    return vehicules_disponibles​
+def liste_vehicules_disponibles(categorie, agence, reservation_date_depart, reservation_date_retour):​
+    liste_contrats_conclus_bloquants_sur_la_periode_demandee = (​
+        Contrat.objects.all() # On part de la liste de tous les contrats puis on va la filtrer​
+        .filter(agence=agence) # on retient ceux signés dans l'agence concernée par la recherche​
+        .filter(date_fin__gte=reservation_date_depart) # si la date de retour d'un contrat existant est postérieur à notre date de départ​
+        .filter(date_debut__lte=reservation_date_retour) # et si la date de début d'un contrat existant est antérieur à notre date de retour​
+    )​
+​
+    vehicules_disponibles = (​
+        Vehicule.objects.all() # On part de la liste de tous les véhicules puis on va la filtrer​
+        #.filter(active=True) # à condition qu'ils soient actifs​
+        .filter(categorie=categorie) # à condition que la catégorie corresponde à celle demandée dans le formulaire​
+        .filter(agence=agence)  #à condition que l'agence corresponde à celle demandée dans le formulaire​
+        .exclude(id__in=Subquery(liste_contrats_conclus_bloquants_sur_la_periode_demandee.values('vehicule'))) #à condition que le véhicule ne figure pas dans un contrat déjà conclu aux dates souhaitées​
+    )​
+    return vehicules_disponibles​
+'''
+
+'''
+# Fonction de sélection de véhicule disponibles
+# (appelée à partir du moment où le formulaire agence / dates de réservations) est valide
+
+# En fonction des données soumies et de leue validité, affiche un template ou un autre :
+#       confirm-reservation-infos-vehicule.html
+#  ou   enregistrement-contrat.html
+def selection_vehicule(request, reservation, reservation_form):
+    #Récupération des informations soumies par le formualire
+    categorie = reservation.get('categorie')​
+    agence = reservation.get('agence_depart')​
+    date_depart = reservation.get('date_depart')​
+    date_retour = reservation.get('date_retour')​
+    vehicule_id = request.POST.get('vehicule_id')​
+    reservation_step = request.POST.get('reservation-step')​
+​
+    if (vehicule_id):​
+        client_form = ClientForm()​
+        if (request.user.client):​
+            client_form = ClientForm(instance=request.user.client)​
+        if (reservation_step and reservation_step == 'infos-client'):​
+            client_form = ClientForm(request.POST, instance=request.user.client or None)​
+            if (client_form.is_valid()):​
+                client = client_form.save(commit=False)​
+                if (not client.user):​
+                    client.user = request.user​
+                client.save()​
+​
+                vehicule = Vehicule.objects.get(id=vehicule_id)​
+                already_contracted = (​
+                    Contrat.objects.all()​
+                        .filter(agence=agence)​
+                        .filter(vehicule=vehicule)​
+                        .filter(date_fin__gte=date_depart)​
+                        .filter(date_debut__lte=date_retour)​
+                )​
+​
+                error = ''​
+                if already_contracted.count() > 0:​
+                    ctr_error = 'Impossible de réserver à partir des choix transmis. Le véhicule demandé n\'est plus disponible.'​
+                    context = {'ctr_error': ctr_error}​
+                    return render(request, 'enregistrement-contrat.html', context)​
+                elif not error:​
+                    # création du contrat​
+                    ctr_error = ''​
+                    contrat = Contrat(vehicule=vehicule, client=client, agence=agence, date_debut=date_depart, date_fin=date_retour)​
+                    contrat.save()​
+​
+                    # context​
+                    context = {​
+                        'ctr_error': ctr_error,​
+                        'agence': agence,​
+                        'vehicule': vehicule,​
+                        'date_debut': date_depart,​
+                        'date_fin': date_retour,​
+                        'client': client,​
+                        'contrat': contrat,​
+                    }​
+                    return render(request, 'enregistrement-contrat.html', context)​
+                elif(error):​
+                    ctr_error = error​
+                    context = {'ctr_error': ctr_error}​
+                    return render(request, 'enregistrement-contrat.html', context)​
+​
+        context = {'vehicule_id': vehicule_id, 'reservation': reservation, 'client_form': client_form}​
+        return render(request, 'confirm-reservation-infos-client.html', context)​
+​
+    #Récupération de la liste des véhicules disponibles​
+    vehicules = liste_vehicules_disponibles(categorie, agence, date_depart, date_retour)​
+​
+    context = {'vehicules': vehicules, 'reservation': reservation, 'reservation_form': reservation_form}​
+    return render(request, 'confirm-reservation-infos-vehicule.html', context)​
+'''
