+++ superloc/visitor/static/visitor/style.css
@@ -315,7 +315,7 @@ figcaption
     flex-direction: row;
 }
 
-#start
+#departure
 {
     display: flex;
     flex-direction: column;
