+++ superloc/visitor/templates/visitor/index.html
@@ -68,16 +68,16 @@
                         </select>
                     </div>
                     <div id ="date_agency">
-                        <div id="start">
+                        <div id="departure">
                             <div class="date">
-                                <label for="date_start">Départ</label><br>
-                                <input type="datetime-local" id="date_start"
+                                <label for="date_departure">Départ</label><br>
+                                <input type="datetime-local" id="date_departure"
                                 name="date_start"
-                                min="2021-10-01T00:00" max="2022-12-01T00:00" required>
+                                min="2022-01-01T00:00" max="2022-12-01T00:00" required>
                             </div>
                             <div class="agency">
                                 Agence<br>
-                                <select id="agency_start" required>
+                                <select id="agency_departure" required>
                                     <option value="1">Paris - Dauphine</option>
                                 </select>
                             </div>
@@ -87,7 +87,7 @@
                                 <label for="date_back">Retour</label><br>
                                 <input type="datetime-local" id="date_back"
                                 name="date_back"
-                                min="2021-10-01T00:00" max="2022-12-01T00:00" required>
+                                min="2022-01-01T00:00" max="2022-12-01T00:00" required>
                             </div>
                             <div class="agency">
                                 Agence<br>
